+++ src/common/layouts/AuthLayout/hooks/useHeaderDescription.ts
@@ -1,4 +1,4 @@
-import getHeaderDescription from "@/common/components/AuthLayout/utils/getHeaderDescription";
+import getHeaderDescription from "@/common/layouts/AuthLayout/utils/getHeaderDescription";
 import { useLocation } from "react-router-dom";
 
 export default function useHeaderDescription() {
