+++ src/common/layouts/Layout/Footer/index.tsx
@@ -1,6 +1,6 @@
-import Copyright from "@/common/components/Layout/Footer/Copyright";
-import ProjectLink from "@/common/components/Layout/Footer/GithubLink";
 import { teamMembers } from "@/common/constants";
+import Copyright from "@/common/layouts/Layout/Footer/Copyright";
+import GithubLink from "@/common/layouts/Layout/Footer/GithubLink";
 
 export default function Footer() {
   return (
@@ -8,10 +8,10 @@ export default function Footer() {
       <div className="flex w-content-width flex-col items-center justify-center p-8">
         <div className="mb-4 flex gap-2">
           {teamMembers.map((member) => (
-            <ProjectLink content={member} key={member.name} />
+            <GithubLink content={member} key={member.name} />
           ))}
         </div>
-        <ProjectLink />
+        <GithubLink />
         <Copyright />
       </div>
     </footer>
